Converted the linear solver choice from conditional compilation to an enum and started to add a unit test for the Solver class.

+++ matrix.h
@@ -75,7 +75,7 @@ namespace SYNARMOSMA {
     void increment(unsigned int,unsigned int,kind);
     void convert(kind*) const;
     bool empty_row(unsigned int) const;
-    float sparsity() const;
+    double sparsity() const;
     bool diagonally_dominant() const;
     void multiply(const std::vector<kind>&,std::vector<kind>&) const;
     void get_diagonal(std::vector<kind>&) const;
